Added Skills and it relationships

+++ models/src/main/java/ar/edu/itba/paw/models/Post.java
@@ -6,10 +6,10 @@ public class Post {
 	private Long id;
 	private String title;
 	private String description;
-	private int userId;
+	private Long userId;
 	private Date createdAt;
 
-	public Post(Long id, String title, String description, int userId, Date createdAt) {
+	public Post(Long id, String title, String description, Long userId, Date createdAt) {
 		this.id = id;
 		this.title = title;
 		this.description = description;
@@ -29,7 +29,7 @@ public class Post {
 		return description;
 	}
 
-	public int getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
